daniel/Detectron/mycode/http/requirements.txt
flask
werkzeug
requests
pillow
